Extending Bias, Interpretations and Bug fixes (#4)
* bias extended for 00d62c1b

* added timeout and removed seperator interpretation

* broke it

* yes!

* more optimisations

* WIP

* bug fix

* WIP

* remove Constant

* added interpretatation

* good progress. Fixed bug in Foil

* 1b2d62fb

* that was painful

* wip

* wip

* files

* will try this live

* last commit for this swipe

+++ arc_2024/data_management/data_manager.py
@@ -159,7 +159,10 @@ class DataManager:
                 raise ValueError("Array must be 2D")
 
             solution.append(
-                {"attempt_1": output.tolist(), "attempt_2": output.tolist()}
+                {
+                    "attempt_1": output.astype(int).tolist(),
+                    "attempt_2": output.astype(int).tolist(),
+                }
             )
 
         # Update the JSON content
